Separated game and sfml_game
