+++ djog_unos_2018.pri
@@ -1,5 +1,7 @@
 SOURCES += \
-    game.cpp
+    game.cpp \
+    $$PWD/sfml_game.cpp
 
 HEADERS += \
-    game.h
+    game.h \
+    $$PWD/sfml_game.h
